Rename lib to util
