+++ components/EmailMultiInput/EmailMultiInput.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import TagInput from 'react-tagsinput';
-import { backspace, tab, enter, del, comma } from '../../lib/keyCodes';
+import { backspace, tab, enter, del, comma } from '../../util/keyCodes';
 
 import s from './EmailMultiInput.css';
